Remove TODOs related to use of FIPS endpoints with DynamoDB
Decided to keep the current code as-is, so removing the TODOs

aws/aws-sdk-go#5078 was landed upstream, so bump `aws-sdk-go` to pick up the change.
Note that this new functionality is not used, but would like to keep it updated
so it can possibly be used in the future without needing to bump the SDK version.